LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/block-view-timeline-current-time.tentative-expected.txt


FAIL View timeline with start and end scroll offsets that do not align with the scroll boundaries assert_approx_equals: values do not match for "Timeline's currentTime at container start boundary" expected -150 +/- 0.125 but got -152.5
FAIL View timeline does not clamp starting scroll offset at 0 assert_approx_equals: values do not match for "Timeline's currentTime at container start boundary" expected 50 +/- 0.125 but got 47.5
FAIL View timeline  does not clamp end scroll offset at max scroll assert_approx_equals: values do not match for "Timeline's currentTime at container start boundary" expected -150 +/- 0.125 but got -152.5
